+++ packages/GoBarber/backend/src/app/controllers/UserController.js
@@ -1,23 +1,8 @@
-import * as Yup from 'yup';
 import User from '../models/User';
 import File from '../models/File';
 
 class UserController {
   async store(req, res) {
-    const schema = Yup.object().shape({
-      name: Yup.string().required(),
-      email: Yup.string()
-        .email()
-        .required(),
-      password: Yup.string()
-        .required()
-        .min(6),
-    });
-
-    if (!(await schema.isValid(req.body))) {
-      return res.status(400).json({ error: 'validations fails' });
-    }
-
     const userExists = await User.findOne({ where: { email: req.body.email } });
 
     if (userExists) {
@@ -34,24 +19,6 @@ class UserController {
   }
 
   async update(req, res) {
-    const schema = Yup.object().shape({
-      name: Yup.string(),
-      email: Yup.string().email(),
-      oldPassword: Yup.string().min(6),
-      password: Yup.string()
-        .min(6)
-        .when('oldPassword', (oldPassword, field) =>
-          oldPassword ? field.required() : field
-        ),
-      confirmPassword: Yup.string().when('password', (password, field) =>
-        password ? field.required().oneOf([Yup.ref('password')]) : field
-      ),
-    });
-
-    if (!(await schema.isValid(req.body))) {
-      return res.status(400).json({ error: 'validations fails' });
-    }
-
     const { email, oldPassword } = req.body;
     const user = await User.findByPk(req.userId);
