+++ packages/GoBarber/backend/src/app/validators/AppointmentStore.js
@@ -1,0 +1,17 @@
+import * as Yup from 'yup';
+
+export default async (req, res, next) => {
+  try {
+    const schema = Yup.object().shape({
+      providerId: Yup.number().required(),
+      date: Yup.date().required(),
+    });
+
+    await schema.validate(req.body, { abortEarly: false });
+    return next();
+  } catch (err) {
+    return res
+      .status(400)
+      .json({ error: 'validations fails', messages: err.inner });
+  }
+};
