+++ packages/GoBarber/backend/src/app/validators/SessionStore.js
@@ -1,0 +1,19 @@
+import * as Yup from 'yup';
+
+export default async (req, res, next) => {
+  try {
+    const schema = Yup.object().shape({
+      email: Yup.string()
+        .email()
+        .required(),
+      password: Yup.string().required(),
+    });
+
+    await schema.validate(req.body, { abortEarly: false });
+    return next();
+  } catch (err) {
+    return res
+      .status(400)
+      .json({ error: 'validations fails', messages: err.inner });
+  }
+};
