+++ packages/GoBarber/backend/src/app/validators/UserStore.js
@@ -1,0 +1,22 @@
+import * as Yup from 'yup';
+
+export default async (req, res, next) => {
+  try {
+    const schema = Yup.object().shape({
+      name: Yup.string().required(),
+      email: Yup.string()
+        .email()
+        .required(),
+      password: Yup.string()
+        .required()
+        .min(6),
+    });
+
+    await schema.validate(req.body, { abortEarly: false });
+    return next();
+  } catch (err) {
+    return res
+      .status(400)
+      .json({ error: 'validations fails', messages: err.inner });
+  }
+};
